tech: Remove MockServerError and move TransferProtocol
This PR removes the old MockServerError and moves (nests)
the TransferProtocol type into the MockServer
